Update hufflepuff
Create a for loop that iterates over an array. On the first iteration the loop will log to the console the the fist element in the array and the last element. On the second iteration the loop will log to the console the 2nd element in the array and the second to the last element. This process will continue until all the element in the array are logged.

This is how my code is itirating. It is repeating in reverse. 

'I'
' '
'L'
'o'
'v'
'e'
' '
'C'
'o'
'd'
'e'
's'
'm'
'i'
't'
'h'
'h'
't'
'i'
'm'
's'
'e'
'd'
'o'
'C'
' '
'e'
'v'
'o'
'L'
' '
'I'